requirements.txt
asgiref==3.9.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.5
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
